requirements.txt
langchain==0.3.23
langchain-anthropic==0.3.10
anthropic==0.49.0
langchain-core==0.3.51
pydantic==2.10.6
python-dotenv==1.1.0 
diplomacy==0.1.0
openai==1.1.0